DiffUtils

DiffUtils.diff

DiffUtils.diff

DiffUtils.diff — Method

DiffUtils.diff(x1, x2; stdout=stdout, side_by_side=true, color=true, options...)

Prints the diff of the string representations of x1 and x2.

Optionally, stdout specifies the stream which the output of diff gets written to.

julia> DiffUtils.diff("foo", "bar")
foo							      |	bar

For further details on the supported options, see diff(::Function).

DiffUtils.diff — Method

DiffUtils.diff(f::Function; stdout=stdout, side_by_side=true, color=true, options...)

Takes a function f which gets passed two output streams and may write to them. The resulting contents are then passed to diff.

Optionally, stdout specifies the stream which the output of diff gets written to.

julia> DiffUtils.diff() do s1, s2
           println(s1, "foo")
           println(s2, "bar")
       end
foo							      |	bar

Supported Options

See the diff man page for further reference.

Boolean Arguments

text, brief, recursive, unidirectional_new_file, ignore_all_space, d, ignore_blank_lines, side_by_side, ignore_case, y, ignore_file_name_case, p, e, left_column, ed, Z, N, s, paginate, initial_tab, show_c_function, suppress_blank_empty, suppress_common_lines, ignore_tab_expansion, minimal, strip_trailing_cr, expand_tabs, no_ignore_file_name_case, new_file, r, T, a, i, b, n, speed_large_files, report_identical_files, no_dereference, B, l, w, E, q, ignore_space_change, rcs, ignore_trailing_space, t

Integer Arguments

U, width, horizon_lines, context, W, tabsize, c, u, unified, C

Other

color, labels